ordiggplot scales biplot scores to match all other scores
There seems to be no way of scaling arrows in geom_ calls, as these
do not know about other layers in final plot. As a crude way we now
multiply with scaling factor in the beginning. If row and species
scores are scaled very differently (like is common in RDA), the
final scaling can be add, and we may need to find ways to circumvent
problems in _arrow function (which does not yet exist).