export vectorfit for easier testing (but yet undocumented)
Currently stat_vectorfit is unfacetable and ungroupable: if
the function receives a subset of data (facet_, group=), an error
is generated. The 'data' will be subsetted, but not 'edata'. Even
if we fix this in code, ggplot2 will alert an error of non-matching
data. The data and edata are first evaluated when the ggplot is built
up, and so there is no non-hackery way of dealing with this (the
'data' will not be a data.frame but a function until building the
final ggplot in print.ggplot after the command chain).